Moved controller extension method to service
